+++ CHANGELOG.md
@@ -128,4 +128,7 @@
 * Improved attributions functionality. Now extracts the data from the layers and not form the ui control
 * Count loaded tiles and images in all layers (not only wms)
 * Watermark logo now accepts non square images
+* Added watermark logo in the example
+* Added controls (zoom and rotation) on the reframe instance
 * Improved css style in prinFields using flex
+* Updated screenshots